fix: convert data to bigdecimal

+++ src/main/java/dev/braianz/DesafioItau/Models/Estatistica.java
@@ -4,14 +4,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Estatistica {
     private long count;
-    private double sum;
-    private double avg;
-    private double min;
-    private double max;
+    private BigDecimal sum;
+    private BigDecimal avg;
+    private BigDecimal min;
+    private BigDecimal max;
 
 }
